# docs/api/openapi.yml
openapi: "3.0.3"
info:
  title: OpenBench API
  description: |
    Repair shop management system — ticket tracking, inventory, POS sales, and warranty management.

    Base URL: `http://localhost:8080` (dev) | `https://api.openbench.example.com` (prod)

    ## Authentication

    Auth uses HTTP-only cookies (`access_token`, `refresh_token`) with Token Rotation (RTR).
    Protected endpoints require `credentials: 'include'` on the client side.

    | Cookie | Type | Description |
    |--------|------|-------------|
    | `access_token` | JWT (short-lived) | Sent automatically on credentialed requests |
    | `refresh_token` | Opaque (long-lived) | Rotated on each `/auth/refresh` call |

    ## Response Envelope

    Every response follows this shape:

    ```json
    {
      "code": 200,
      "message": "Success",
      "data": { ... }
    }
    ```

    Error responses omit `data` and include an `error` field with a technical message.
  version: "0.1.0"
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.openbench.example.com
    description: Production server
paths:
...
  /api/v1/admin/inventory:
    get:
      operationId: listInventory
      summary: List all products
      description: Returns all inventory products — both retail items and spare parts.
      tags: [Inventory]
      security:
        - CookieAuth: []

      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ProductList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createProduct
      summary: Create a product
      description: Adds a new product to inventory.
      tags: [Inventory]
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ProductData"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/inventory/{id}:
    get:
      operationId: getProduct
      summary: Get a single product
      description: Returns a product by its ID.
      tags: [Inventory]
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product UUID
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ProductData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Error"
    patch:
      operationId: updateProduct
      summary: Update a product
      description: Updates one or more fields of a product (stock, price, etc.).
      tags: [Inventory]
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_ProductData"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Error"
    delete:
      operationId: deleteProduct
      summary: Delete a product
      description: Removes a product from the inventory catalog.
      tags: [Inventory]
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product UUID
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Empty"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Error"

  # ────────────────────────────── SALES ───────────────────────────────
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token set via Set-Cookie on sign-in / refresh.
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse_Error"
    Forbidden:
      description: Insufficient permissions (role must be admin)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse_Error"
  schemas:
...
    ApiResponse_Empty:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "Success"
        data:
          nullable: true
          example: null
      required: [code]
    ApiResponse_Error:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: "Invalid email or password"
        error:
          type: string
          example: "invalid credentials"
      required: [code, message]
...
    ApiResponse_ProductList:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "Inventory retrieved"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [code, message, data]
    ApiResponse_ProductData:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "Product retrieved"
        data:
          $ref: "#/components/schemas/Product"
      required: [code, message, data]
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-..."
        name:
          type: string
          example: "Tempered Glass Samsung S24"
        category:
          type: string
          enum: [retail, spare_part]
          example: spare_part
        stock:
          type: integer
          example: 50
        price:
          type: number
          format: float
          example: 50000
        cost_price:
          type: number
          format: float
          example: 25000
        min_stock:
          type: integer
          example: 10
      required: [id, name, category, stock, price, cost_price, min_stock]
    ProductCreate:
      type: object
      required: [name, category, stock, price, cost_price, min_stock]
      properties:
        name:
          type: string
          example: "Tempered Glass Samsung S24"
          maxLength: 255
          x-oapi-codegen-extra-tags:
            validate: "required,max=255"
        category:
          type: string
          enum: [retail, spare_part]
          example: spare_part
          x-oapi-codegen-extra-tags:
            validate: "required,oneof=retail spare_part"
        stock:
          type: integer
          minimum: 0
          example: 50
          x-oapi-codegen-extra-tags:
            validate: "gte=0"
        price:
          type: number
          format: float
          minimum: 0
          example: 50000
          x-oapi-codegen-extra-tags:
            validate: "gte=0"
        cost_price:
          type: number
          format: float
          minimum: 0
          example: 25000
          x-oapi-codegen-extra-tags:
            validate: "gte=0"
        min_stock:
          type: integer
          minimum: 0
          example: 10
          x-oapi-codegen-extra-tags:
            validate: "gte=0"
      additionalProperties: false
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Tempered Glass Samsung S24 Ultra"
          maxLength: 255
          x-oapi-codegen-extra-tags:
            validate: "omitempty,min=1,max=255"
        category:
          type: string
          enum: [retail, spare_part]
          example: spare_part
          x-oapi-codegen-extra-tags:
            validate: "omitempty,oneof=retail spare_part"
        stock:
          type: integer
          minimum: 0
          example: 45
          x-oapi-codegen-extra-tags:
            validate: "omitempty,gte=0"
        price:
          type: number
          format: float
          minimum: 0
          example: 55000
          x-oapi-codegen-extra-tags:
            validate: "omitempty,gte=0"
        cost_price:
          type: number
          format: float
          minimum: 0
          example: 30000
          x-oapi-codegen-extra-tags:
            validate: "omitempty,gte=0"
        min_stock:
          type: integer
          minimum: 0
          example: 10
          x-oapi-codegen-extra-tags:
            validate: "omitempty,gte=0"
      additionalProperties: false

    # ───────────── SALE ─────────────
...